+++ CHANGELOG.md
@@ -3,6 +3,12 @@ Release Notes
 
 ## Next Version
 
+### New
+
+- `Database.isSQLiteInternalTable(_:)` returns whether a table name is an internal SQLite table.
+- `Database.isGRDBInternalTable(_:)` returns whether a table name is an internal GRDB table.
+- `DatabaseMigrator.appliedMigrations(in:)` returns the set of applied migrations identifiers in a database.
+
 ### Breaking Changes
 
 - The Record protocols have been renamed: `RowConvertible` to `FetchableRecord`, `Persistable` to `PersistableRecord`, and `TableMapping` to `TableRecord`.